# GoPro/hw_3/openapi.yaml
openapi: 3.0.3
info:
  title: Tasks API
  description: |
    REST API для управления списком задач (to-do) в оперативной памяти.

    **Особенности реализации:**
    * Все операции потокобезопасны и детерминированы (список всегда отсортирован по ID).
    * Поле `title` проходит сквозную нормализацию (trimming). В систему сохраняется и возвращается значение без пробелов по краям.
    * Неподдерживаемые методы строго возвращают код `405` и заголовок `Allow`.
  version: 1.2.0
paths:
...
  /tasks:
    get:
      summary: Получить список всех задач (отсортирован по ID)
      responses:
        '200':
          description: Успешное получение списка задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '405':
          $ref: '#/components/responses/405MethodNotAllowedCollection'
    post:
      summary: Создать новую задачу
      description: Пробелы по краям поля `title` автоматически удаляются перед валидацией и сохранением.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: Забрать заказ из Озона
                done:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Задача успешно создана (возвращается нормализованное значение)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '405':
          $ref: '#/components/responses/405MethodNotAllowedCollection'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    put: &collection405
      summary: Неподдерживаемый метод для коллекции /tasks
      deprecated: true
      responses:
        '405':
          $ref: '#/components/responses/405MethodNotAllowedCollection'
    delete: *collection405
    patch: *collection405
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Купить молоко
        done:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-08-06T20:20:15+03:00"
    Error:
      type: object
      properties:
        error:
          type: string
          example: Описание ошибки
  responses:
    400BadRequest:
      description: Неверные входные данные, пустой title или невалидный формат JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    500InternalServerError:
      description: Внутренний непредвиденный сбой сервера или хранилища
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    405MethodNotAllowedCollection:
      description: Метод не поддерживается для коллекции /tasks
      headers:
        Allow:
          schema:
            type: string
            example: GET, POST
          description: Список разрешенных методов
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...